+++ ARL/app/services/__init__.py
@@ -32,6 +32,7 @@ from .trufflehog_scan import run_trufflehog_js
 from .afrog_scan import run_afrog_scan
 from .baseUpdateTask import BaseUpdateTask
 from .domainSiteUpdate import domain_site_update
-from .fingerprint_cache import finger_db_cache, finger_db_identify, have_human_rule_from_db
+from .fingerprint_cache import finger_db_cache, finger_db_identify, finger_db_identify_detail, \
+    have_human_rule_from_db
 from .fingerprint import FingerPrint
 from .expr import evaluate_expression, check_expression, check_expression_with_error